// src/types/drone-env.ts
export interface DroneEnv {
  CI: boolean
  DRONE: boolean
  DRONE_BRANCH: string
  DRONE_BUILD_ACTION: string
  DRONE_BUILD_CREATED: number
  DRONE_BUILD_EVENT: string
  DRONE_BUILD_FINISHED: number
  DRONE_BUILD_NUMBER: number
  DRONE_BUILD_PARENT: number
  DRONE_BUILD_STARTED: number
  DRONE_BUILD_STATUS: string
  DRONE_COMMIT: string
  DRONE_COMMIT_AFTER: string
  DRONE_COMMIT_AUTHOR: string
  DRONE_COMMIT_AUTHOR_AVATAR: string
  DRONE_COMMIT_AUTHOR_EMAIL: string
  DRONE_COMMIT_AUTHOR_NAME: string
  DRONE_COMMIT_BEFORE: string
  DRONE_COMMIT_BRANCH: string
  DRONE_COMMIT_LINK: string
  DRONE_COMMIT_MESSAGE: string
  DRONE_COMMIT_REF: string
  DRONE_COMMIT_SHA: string
  DRONE_DEPLOY_TO: string
  DRONE_FAILED_STAGES: string[]
  DRONE_FAILED_STEPS: string[]
  DRONE_GIT_HTTP_URL: string
  DRONE_GIT_SSH_URL: string
  DRONE_PULL_REQUEST: number
  DRONE_REPO: string
  DRONE_REPO_BRANCH: string
  DRONE_REPO_LINK: string
  DRONE_REPO_NAME: string
  DRONE_REPO_NAMESPACE: string
  DRONE_REPO_OWNER: string
  DRONE_REPO_PRIVATE: boolean
  DRONE_REPO_SCM: 'git' | 'hg' | 'svn'
  DRONE_REPO_VISIBILITY: 'public' | 'private' | 'internal'
  DRONE_SEMVER: string
  DRONE_SEMVER_BUILD: string
  DRONE_SEMVER_ERROR: string
  DRONE_SEMVER_MAJOR: string
  DRONE_SEMVER_MINOR: string
  DRONE_SEMVER_PATCH: string
  DRONE_SEMVER_PRERELEASE: string
  DRONE_SEMVER_SHORT: string
  DRONE_SOURCE_BRANCH: string
  DRONE_STAGE_ARCH: string
  DRONE_STAGE_DEPENDS_ON: string[]
  DRONE_STAGE_FINISHED: number
  DRONE_STAGE_KIND: string
  DRONE_STAGE_MACHINE: string
  DRONE_STAGE_NAME: string
  DRONE_STAGE_NUMBER: string
  DRONE_STAGE_OS:
    | 'darwin'
    | 'dragonfly'
    | 'freebsd'
    | 'linux'
    | 'netbsd'
    | 'openbsd'
    | 'solaris'
    | 'windows'
  DRONE_STAGE_STARTED: number
  DRONE_STAGE_STATUS: string
  DRONE_STAGE_TYPE: string
  DRONE_STAGE_VARIANT: string
  DRONE_STEP_NAME: string
  DRONE_STEP_NUMBER: string
  DRONE_SYSTEM_HOST: string
  DRONE_SYSTEM_HOSTNAME: string
  DRONE_SYSTEM_PROTO: string
  DRONE_SYSTEM_VERSION: string
  DRONE_TAG: string
  DRONE_TARGET_BRANCH: string
}